# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_tmalldedede_agentbox_internal_profile.DebugConfig:
    properties:
      verbose:
        type: boolean
    type: object
  github_com_tmalldedede_agentbox_internal_profile.FeatureConfig:
    properties:
      web_search:
        type: boolean
    type: object
  github_com_tmalldedede_agentbox_internal_profile.MCPServerConfig:
    properties:
      args:
        items:
          type: string
        type: array
      command:
        type: string
      description:
        type: string
      env:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
    type: object
  github_com_tmalldedede_agentbox_internal_profile.ModelConfig:
    properties:
      base_url:
        description: Custom API endpoint (proxy/compatible API)
        type: string
      bearer_token:
        description: experimental_bearer_token for Codex
        type: string
      disable_traffic:
        description: CLAUDE_CODE_DISABLE_NONESSENTIAL_TRAFFIC
        type: boolean
      env_key:
        description: Environment variable name for API key (e.g., "ZHIPU_API_KEY")
        type: string
      haiku_model:
        description: Model tier configuration (Claude Code)
        type: string
      max_output_tokens:
        description: CLAUDE_CODE_MAX_OUTPUT_TOKENS
        type: integer
      name:
        description: Basic configuration
        type: string
      opus_model:
        description: ANTHROPIC_DEFAULT_OPUS_MODEL
        type: string
      provider:
        description: e.g., "anthropic", "openai", "glm"
        type: string
      reasoning_effort:
        description: low/medium/high (Codex)
        type: string
      sonnet_model:
        description: ANTHROPIC_DEFAULT_SONNET_MODEL
        type: string
      timeout_ms:
        description: Advanced configuration
        type: integer
      wire_api:
        description: Codex specific - third-party provider support
        type: string
    type: object
  github_com_tmalldedede_agentbox_internal_profile.PermissionConfig:
    properties:
      additional_dirs:
        description: Common
        items:
          type: string
        type: array
      allowed_tools:
        description: --allowedTools
        items:
          type: string
        type: array
      approval_policy:
        description: untrusted/on-failure/on-request/never
        type: string
      disallowed_tools:
        description: --disallowedTools
        items:
          type: string
        type: array
      full_auto:
        description: --full-auto
        type: boolean
      mode:
        description: Claude Code specific
        type: string
      sandbox_mode:
        description: Codex specific
        type: string
      skip_all:
        description: --dangerously-skip-permissions
        type: boolean
      tools:
        description: --tools
        items:
          type: string
        type: array
    type: object
  github_com_tmalldedede_agentbox_internal_profile.Profile:
    properties:
      adapter:
        description: 'Adapter selection: "claude-code" | "codex"'
        type: string
      append_system_prompt:
        type: string
      base_instructions:
        type: string
      config_overrides:
        additionalProperties:
          type: string
        description: Config overrides (Codex specific)
        type: object
      created_at:
        description: Metadata
        type: string
      created_by:
        type: string
      credential_id:
        description: Credential ID for API authentication
        type: string
      custom_agents:
        description: Custom agents (Claude Code specific)
        items:
          type: integer
        type: array
      debug:
        allOf:
        - $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_profile.DebugConfig'
        description: Debug settings
      description:
        type: string
      developer_instructions:
        type: string
      extends:
        description: 'Inheritance: extends another Profile'
        type: string
      features:
        allOf:
        - $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_profile.FeatureConfig'
        description: Feature flags
      icon:
        type: string
      id:
        description: Basic info
        type: string
      is_built_in:
        type: boolean
      is_public:
        type: boolean
      mcp_servers:
        description: MCP servers
        items:
          $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_profile.MCPServerConfig'
        type: array
      model:
        allOf:
        - $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_profile.ModelConfig'
        description: Model configuration
      name:
        type: string
      output_format:
        description: Output configuration
        type: string
      output_schema:
        description: JSON Schema file path
        type: string
      permissions:
        allOf:
        - $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_profile.PermissionConfig'
        description: Permission configuration
      resources:
        allOf:
        - $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_profile.ResourceConfig'
        description: Resource limits
      system_prompt:
        description: System prompts
        type: string
      tags:
        items:
          type: string
        type: array
      updated_at:
        type: string
    type: object
  github_com_tmalldedede_agentbox_internal_profile.ResourceConfig:
    properties:
      cpus:
        description: Container resources
        type: number
      disk_gb:
        type: integer
      max_budget_usd:
        type: number
      max_tokens:
        type: integer
      max_turns:
        type: integer
      memory_mb:
        type: integer
      timeout:
        $ref: '#/definitions/time.Duration'
    type: object
...
  internal_api.CloneRequest:
    properties:
      new_id:
        type: string
      new_name:
        type: string
    required:
    - new_id
    - new_name
    type: object
  internal_api.CreateProfileRequest:
    properties:
      adapter:
        type: string
      append_system_prompt:
        type: string
      base_instructions:
        type: string
      config_overrides:
        additionalProperties:
          type: string
        type: object
      credential_id:
        type: string
      debug:
        $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_profile.DebugConfig'
      description:
        type: string
      developer_instructions:
        type: string
      extends:
        type: string
      features:
        $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_profile.FeatureConfig'
      icon:
        type: string
      id:
        type: string
      is_public:
        type: boolean
      mcp_servers:
        items:
          $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_profile.MCPServerConfig'
        type: array
      model:
        $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_profile.ModelConfig'
      name:
        type: string
      output_format:
        type: string
      output_schema:
        type: string
      permissions:
        $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_profile.PermissionConfig'
      resources:
        $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_profile.ResourceConfig'
      system_prompt:
        type: string
      tags:
        items:
          type: string
        type: array
    required:
    - adapter
    - id
    - name
    type: object
...
  internal_api.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      type:
        description: 错误类型
        type: string
    type: object
...
  time.Duration:
    enum:
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    format: int64
    type: integer
    x-enum-varnames:
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
host: localhost:18080
info:
  contact:
    email: support@agentbox.dev
    name: AgentBox Team
    url: https://github.com/user/agentbox
  description: |-
    AgentBox - Open-source AI Agent containerized runtime platform

    ## Overview
    AgentBox provides a unified API for managing AI coding agents (Claude Code, Codex, OpenCode) in containerized environments.

    ## Authentication
    Currently no authentication required. API keys are managed via Credentials API.

    ## API Structure
    - **Public API** (`/api/v1/*`): Core functionality for external integrations
    - **Admin API** (`/api/v1/admin/*`): Platform management operations
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: AgentBox API
  version: "1.0"
paths:
...
  /profiles:
    get:
      description: Get a list of all available runtime configuration profiles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_profile.Profile'
                  type: array
              type: object
      summary: List all profiles
      tags:
      - Profiles
    post:
      consumes:
      - application/json
      description: Create a new runtime configuration profile for agents
      parameters:
      - description: Profile configuration
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/internal_api.CreateProfileRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_profile.Profile'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Create a profile
      tags:
      - Profiles
  /profiles/{id}:
    delete:
      description: Delete an existing profile (built-in profiles cannot be deleted)
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Delete a profile
      tags:
      - Profiles
    get:
      description: Get detailed information about a specific profile
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_profile.Profile'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Get a profile
      tags:
      - Profiles
    put:
      consumes:
      - application/json
      description: Update an existing profile configuration
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      - description: Profile configuration
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/internal_api.CreateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_profile.Profile'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Update a profile
      tags:
      - Profiles
  /profiles/{id}/clone:
    post:
      consumes:
      - application/json
      description: Create a copy of an existing profile with a new ID and name
      parameters:
      - description: Source Profile ID
        in: path
        name: id
        required: true
        type: string
      - description: Clone configuration
        in: body
        name: clone
        required: true
        schema:
          $ref: '#/definitions/internal_api.CloneRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_profile.Profile'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Clone a profile
      tags:
      - Profiles
  /profiles/{id}/resolved:
    get:
      description: Get a profile with inheritance chain resolved (merges parent profile
        settings)
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_profile.Profile'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Get resolved profile
      tags:
      - Profiles
...
swagger: "2.0"
...